+++ lib/properties_intf.ml
@@ -1339,7 +1339,7 @@ type text_box_edge =
   | Var of text_box_edge var
 
 type text_box =
-  | Box of text_box_trim * text_box_edge option
+  | Box of text_box_trim option * text_box_edge option
   | Inherit
   | Initial
   | Unset